refact/#7: clean up base parser
move methods regarnding missing entities into new script
